Set Game
SET is a board game where any table becomes the board. It is a highly
addictive, original game of visual perception. The goal is to be the
first to see a SET from the 12 cards laid face up on the table. A SET
is three cards that are either all the same or all different in each
individual feature. With no turns and no luck, SET is challenging, fast
and fun!